# api-gateway/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "PMS-SIG",
        "description": "Pallet Management System - SIG",
        "license": {
            "name": "Semen Indonesia Group",
            "url": "https://sig.id"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "API Auth ( Register )",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Data untuk otentikasi",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/postRegister"
                        }
                      }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SigResponsePostSuccess"
                        }
                    },
                    "400": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SigResponsePostFailed"
                        }
                    }
                }
            }
        },
        "/users/login": { 
            "post": {
                "tags": [
                    "Auth"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "API Auth ( Login )",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Data untuk login",
                        "required": true,
                        "schema": {
                          "$ref": "#/definitions/postLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseLoginSuccess"
                        }
                    },
                    "400": {
                        "description": "Error request",
                        "schema": {
                            "$ref": "#/definitions/responseLoginBadRequest"
                        }
                    },
                    "422": {
                        "description": "Error request",
                        "schema": {
                            "$ref": "#/definitions/responseLoginIsEmpty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "postRegister": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "default": "jefry"
              },
              "password": {
                "type": "string",
                "default": "password"
              },
              "email": {
                "type": "string",
                "default": "jefry@gmail.com"
              },
              "fullname": {
                "type": "string",
                "default": "jefry jaya"
              }
            },
            "required": ["username", "password", "email", "fullname"]
        },
...
        "postLogin": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "default": "jefry"
                },
                "password": {
                    "type": "string",
                    "default": "password"
                }
            },
            "required": ["username", "password"]
        },
        "responseLoginSuccess": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "token": {
                            "type": "string"
                        },
                        "refresh_token": {
                            "type": "string"
                        },
                        "data": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                },
                                "fullname": {
                                    "type": "string"
                                },
                                "initials": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": []
                                },
                                "ip": {
                                    "type": "string"
                                },
                                "device": {
                                    "type": "string"
                                },
                                "time_access": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "responseLoginBadRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responseLoginIsEmpty": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "SigResponsePostSuccess": {
            "properties": {
                "posted": {
                    "type": "string"
                }
            }
        },
...
        "SigResponsePostFailed": {
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}